requirements.txt
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.1.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
Django==5.1.2
django-cors-headers==4.6.0
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.19.11
django-utils-six==2.0
djangorestframework==3.14.0
executing==2.1.0
goto-statement==1.2
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
npm==0.1.1
numpy==2.1.1
optional-django==0.1.0
packaging==24.1
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.9
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-gettext==5.0
pytz==2023.3.post1
pywin32==308
pyzmq==26.2.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.2
wcwidth==0.2.13
